bootstrap: hello world!
opnsense-bootstrap(8) is a tool that can completely reinstall a
running system in place for a thorough factory reset or to restore
consistency of all the OPNsense files.  It can also wipe the
configuration directory, but won't do that by default.

It will automatically pick up the latest available version and
build a chain of trust by using current package fingerprints -> CA
root certificates -> HTTPS -> OPNsense package fingerprints.

What it will also do is turn a stock FreeBSD 10.1-RELEASE into
an OPNsense installation, given that UFS was used to install the
root file system.

This is as close as we get for debating "distribution as a package"
in FreeBSD at the moment.  Ultimately, no more kernel patches would
simplify the procedure.  Package manager support for base and kernel
will make this even more simple.  FreeBSD 11 is what we are aiming
for for these reasons.  And we'll definitely stick around.

But, wait, there's more...

What it will likely do at some point in the future is moving an
installation from FreeBSD/OPNsense to HardenedBSD/OPNsense or back.
The HardenedBSD team is doing great work and we don't see a reason
why anyone should be locked out of an eco system just for choosing
a different tool.  Above all, we value choice.

If this tool is a candidate for the FreeBSD ports collection is a
different question to be asked.  For now I don't think the timing
and the scope of only transforming 10.1-RELEASE fits.  But as we
work on this and test more, we can gradually lift the restrictions.

"These go to eleven."  :)

Cheers,
Franco