update: annotate a little more clearly how new expansion works
The update sets now reside in ABI-safe directories, which is
good because we can have multiple sets with different signatures
for different versions if necessary, but also a bit complicated
to follow if a set is not where it should be or missing.

We're certaintly not going to add another override here now
without seeing how this works in practice, mostly because
this was never possible before and I "hid" an internal testing
update by mere mistake.