update: kernel debug support
This simply clears the default debug location so that it's possible
to install a debug kernel without lingering debug objects from older
versions.

The idea is to provide a full debug set as an interchangable kernel,
that way we can make it optional without having to provide another
set and simply reuse the old one with a slightly different name.

Good thing we added a persistent marker file in the set itself, we
can now probe for directory /usr/lib/debug/boot/kernel to see if we
have a debug kernel or not, otherwise it walks like a duck and
certainly quacks like one.  ;)