patch: defer error, patch as much as we can; closes #31
The underlying issue is the patch does not apply cleanly, but we should
rather do the post-processing than skip immediately as some of the damage
is already done.

Ideally everything outside "src/" should be discarded, but that requires
preprocessing of the patch, which we can do when we have nothing better
to do.  :)