verify: hack around multi-repo unfriendlyness in the inherited code; closes #52
This bootstrap code flattens "FINGERPRINTS" directory string so only the
last one was loaded and verified.  While this works now by representing
the directories in a list instead I'm not sure that all the things are
considered here:

o Is the repository even enabled?
o Does the repository have a signature type fingerprint?
o Mixed pub key and fingerprint environments ignore pub keys!

The problems already existed in the code base and are only of auxiliary
interest for the time being.