revert: take more care with automatic flag restoration
When users install from ports tree the automatic flag will be
clobbered and the health check will raise a flag.  Let this
utility repair the health check expectation.

We do this for all files instead of just for direct core
dependencies in order to ensure a clean environment without
lingering packages.  The downside is that opnsense-revert
doesn't work when manually fidding with pkg or ports, but
that is not the scope of this tool anyway.

The vital flag comes from package metadata and restores itself
correctly without further interaction.

PR: https://github.com/opnsense/core/issues/3473