update: recover from fatal operations
We basically never know if pkg will adhere to
keeping the core package around.  Instead, look
for the passing of the configuration file so
that we can assume the worst.

To get through this remember the configurations
in effect prior to the upgrade to get the upsteam
link back and install the top package and all of
its dependencies before returning to the normal
code path.