update: fix behaviour when -t target does not exist
Vital flag was left unset, because pkg-fetch does not error
when it can't fetch a package that does not exist and pkg-install
will not reinstall (even with -f) the package that is still
installed leaving the vital flag unset by proxy.

To be really really sure suppress the potential error of
pkg-install to always reach the pkg-set  spot below.