Update opnsense-bootstrap.sh
I was trying to use this with GCP's FreeBSD 12.2 image. One thing it still likes to do is strip out google-cloud-sdk, and nano if that's previously installed. In the meanwhile, I used shellcheck to audit it for potential issues & applied recommended fixes.

One other thing to look at with this: in virtualized environments like GCP, Azure, etc; the IP address for all interfaces are already assigned, so everything needs to be DHCP. I feel like this script could help fix that ahead of time, but not sure how.

Thanks for making a good product though. Been using it for stuff for years now.