update: fiddle with mirror manipulation #5144
Change order in a way that it makes the most sense (from left to right)
but do the ABI replace afterwards if the mirror directory was not
specified using a directory structure already.  In that case the defaul
ABI is inserted.

Changing the ABI still works using -m "abi/dir" or -A "abi" -m "dir".