update: add clean option for obsolete base files: closes #75
The current logic omits the obsolete base file clean from the major
upgrade sequence to avoid partial upgrade situations as much as
possible.  With exposing an option we could run the cleanup later
in the update sequence or manually still.

While here adjust the "Please reboot" flag for the situation where
either a new kernel was installed or the base system directly updated.
In other cases reboot is handled from the scripting so we don't need
to remind users.