update: read from configuration file
A configuration file has some benefits over magic files.
One open question is if a static configuration is desirable,
but we will have to see working with it.  For now do not
document this behaviour but try it out in the development
version as old and new approach to upgrades are interoperable.