update: move core.lock to pkgs.lock
We do want to create a version file to avoid silly
update loops which are already prevented when using
the kernel/base sets.

The problem is that a packages set does not have a
version so we need to create one but cannot clobber
the "core" version file since that one comes from a
possibly unrelated update.