update: at least remember the last full packages upgrade
When upgrading to prelease with the development version
after all is done and there are no other updates available
the upgrade will try to loop over packages set only since
it has no reference for its version.

This doesn't solve the update path problem that moves the
packages to newer versions no longer matching the packages
set, but at least we can start trusting it to remember the
last set it used for upgrade to not pull it in again.

This also works for subsequent updates where you would see
the packages again (22.1.r1 to 22.1.r2 minor update), but
when done once it's also silent about it.