update: support deferred kernel set install
"-u" will no longer install the kernel directly, but the core
has two hooks for it (one main in upgrade script) and one for
manual recovery on reboot (for scripts that expect it to work
with "-u" and reboot).  The reason we needed to split is to do
pre-upgrade hooks that decide if an upgrade will be performed
or if we need to cancel because the migration scripts fail.

While here fix the "-e" option printing "Nothing to do." and
adjust the "Reboot now." message printing a bit.