opnsense-update: Support unescaped mirror URLs
When running opnsense-update with a custom mirror URL, unescaped
slashes (which are part of every URL) will cause the script to
fail:

    # opnsense-update -ikr whatever -m "http://example.org/"
    sed: 1: "/^[[:space:]]*url:[[:sp ...": bad flag in substitute\
    command: '/'

This patch fixes this by using pipe ('|') instead of slash ('/') as
delimiters in the affected sed command.

Escaped slashes (as in `-m "http:\/\/example.org\/"`) will still work
after applying the patch, so this should not break existing
automation.