update: fix docs on -u
We specifically want -u not to run -K internally even though that
is possible.  We do so because after a round of fetching and staging
the pending updates we want to do last minute checks before proceeding.

We do this also because if we check for free disk space eventually we
want to check the disk space after everthing has been staged to get a
better estimate on the space required.