update: add '-g' for update logs print
The GUI may keep decide to keep an update log in persistent
storage.  Since we don't have the full picture from opnsense-update's
perspective we only read the file if it exists.

That's also to separate concerns and not taint the upgrade log
that shows up in the GUI with most likely irrelevant information.

PR: https://github.com/opnsense/core/issues/9029